Add copy g-cloud-7 draft API endpoint
[#97536644](https://www.pivotaltracker.com/story/show/97536644)

POST /draft-services/<draft_id>/copy creates a copy of the draft
service. List of copied fields is the same as when creating a draft
from existing service, except the service_id is not copied from the
original draft.

Required for the "Make a copy" supplier-frontend button. Instead of
an additional API endpoint it'd be possible to add a new apiclient
method that would GET the original draft from the API and then POST
it to the create draft URL. But in order to do this apiclient would
have to filter all non-data fields from the draft: service id, links,
timestamps etc. API endpoint avoids this by only copying the relevant
model fields from one object to another.

Note: similar to the "make a draft from service" endpoint there's
no validation that the supplier that triggered the request is the
owner of the original service or draft.